Fix duplicate item bug in bs-popover with ng-repeat
If your popover lacks a hasContent() function, Bootstrap will end up calling
the content() function twice, and in the case of bs-popover, this would result
in a redundant $compile.

This doesn't necessarily produce any ill effects, but if your bs-popover
partial contained an ng-repeat, the ng-repeat would be replicated once
for every element in the collection.

For example, this:

  Controller:
  -----------
  $scope.things = [{name: "A"}, {name: "B"}, {name: "C"}];

  Partial:
  --------
  <ul>
     <li ng-repeat="thing in things">{{thing.name}}</li>
  </ul>

  Output:
  -------
   Would yield a popover that contained this:

   - A
   - A
   - A
   - B
   - B
   - B
   - C
   - C
   - C

   But it should have contained this:

   - A
   - B
   - C