Added element id (or element var) to typeahread-updated emit
When emitting the change event on the typeahead add the id of the element or element itself to the emit so the element that was updated can be identified when handling the emit.  This helps when there are multiple typeaheads on a page with different actions when the values are changed.