feat(bsDatepicker): enable attribute-driven language change
If the language changes, the directive should immediately change the language of the datepicker.
Currently, as there's no such method in the datepicker plugin
(see https://github.com/eternicode/bootstrap-datepicker/issues/362), a change is forced by
reinitializing the datepicker itself.
Also, a check has been added to the `regexpMap` function, in order to handle cases when the language
passed to the directive is not present in the datepicker's languages.
The `datepickerSpec` module has been updated accordingly with a test checking the new behaviour.