feat(modal tooltip): Adds a feature to allow for custom name-spacing of the events
Change the use of `prefixClass` to just be used for class prefixes. Added a new option called `prefixEvent` for prefixing the events.
Allows for seperate prefix for events across multiple modals whilst not changing the classes allowing for more customisation.
Fix an error on tipElement that occurs when using the dropdown directive with a dynamically changing object as the source.