fix (tooltip): tooltip's content-template
re:#846 tooltip directive's content-template option throws errors in Firefox 10
The directive gets the content-template then attempts to replace the content div
with the content-template html. The function which does the replacement returns
the modified HTML for the templateEl. The function returns
`templateEl[0].outerHTML` but `outerHTML` is not supported in Firefox until v11.
I have included a shim which handles the case where `el.outerHTML` is not
defined

no breaking changes