fix bsTypeahead: selected object display
bsTypeahead fails back to the model value when it fails to match the selected value to the source. It may happen if the model reference an object and
if the source of candidate for type ahead is always a an array with
new items after each key strokes or select.

The fix uses $parseOptions.displayValue on the modelValue.

Fix for #1260