fix(datepicker): fix duplicate days in picker when DST changes
Depending on the browser timezone, the datepicker might show duplicate days when the DST causes hours to go back - this will happen in São Paulo, Brazil timezone in 18th October 2014 at midnight, when the time will go back 1 hour, so the datepicker will show two 18th days and miss the 19th, as shown in #1133.

If DST causes time to go forward, there is no problem because the day will still be incremented. If time goes back, the fix checks if hours are > 12, meaning we went back to original day, and add +2 hours to compensate for the hours going back, causing the day to increment as intended. (Code copied from jquery ui)