feat(timepicker): add option to floor minutes display to nearest minuteStep whole value
Adds a new option 'roundDisplay' that is checked when the initial time value specified by ngModel is not defined. Previously, the timepicker would use the current time in this case. With this new option, we can tell the picker to round the minutes value so we don't get times like 11:38 or 11:43, but instead get 11:35 or 11:40

Closes #1444