feat (tooltip): add "redraw" event listener
Tooltip now listens for a "redraw" event which signals that the displayed
tooltip should be repositioned. Use cases: When the triggering DOM
element is repositioned while the tooltip is active, the "redraw" event can
be emitted/broadcast and the tooltip will remains adjacent the DOM element.
When the dimensions of a tooltip's template change, the "redraw" event can
be emitted/broadcast and the tooltip will remain centered next to the
triggering DOM element.

This is a possible solution to #1497.