fix(modal): dispose modal child scopes after hidding modal
This is only destroying the modal element before showing another modal. This is to avoid timming issues with the hide animation callback, because we could be showing a new modal before the hide animation callback has been called and then the modal element variables would be replaced with the new modal references.
It's not perfect, but it prevents the scopes from increasing each time a modal is shown, like it was done for tooltip.