Fix for click event not updating the value
I was having an issue with the example on the website using the latest versions of angular-strap and bootstrap. The click event is not fired on the input since it is hidden by the button. The fix listens for the event on the active element which in this case would be the button.