fix(dropdown): options are null if dropdown with inline sibling template  is rendered again after it was previously destroyed
if dropdown has inline sibling template then template is saved to the options during the compile time.
After dropdown is destroyed (for example it was inside the modal) and rendered again (modal reopened) - compile for dropdown is not executed and options are null.
This causes error "angular TypeError: Cannot set property 'scope' of null" and dropdown loses its template.
Solution is to save the template not in the options but in the attributes that are not destroyed.