fix(select): don't attempt to select when tab is pressed and no item is selected
Currently, when tab is pressed to get to a new input from `bs-select`, the `bs-select` will
close and fire `$select.select()`.  If an item isn't select, this will be `undefined` and throw an `Error`.
Now, there is a check that the index of the item to be selected isn't undefined or outside
the bounds of the `$matches`.

closes: #1857