ELIX-17015
Updated dropdown.js to make clicking of the dropdown items more consistent in Safari.  After much digging and trial and error it seems that there was a timing issue in that the drop down hide would fire before the click event had been handled, so simply wrapping hide in a timeout seems to have made things work and doesn't seem to adversely impact other browsers.