fix(collapse): Change key bindings
Revert earlier change around tab and shift keybindings on the keypress
events.

Earlier 508 fixes worked around the ability to keyboard navigate through
this control. Somehow a regression was introduced in that commit that
caused the bs-collapse-toggle to become a keyboard "sink", meaning that
no keyboard action could get away from that focused element.